Fixed linux build: 1. New files to scons. 2. Added idlib to include path of Game (otherwise it is impossible to include smth from idlib with single path on both MSVC/GCC). 3. Convert pointer to int and back via intermediate uintptr_t (this code works only on 32-bit platforms anyway).
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/linuxbuild_2017_06@6967 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/script/Script_Program.h
@@ -643,7 +643,7 @@ ID_INLINE const char *idProgram::GetFilenameForStatement( int index ) {
 //stgatilov #4520: converting between 32-bit offset and 64-bit real address
 ID_INLINE int idProgram::ScriptObjectMemory_Pack(byte *ptr) {
 	if (sizeof(void*) == 4)
-		return (int)ptr;
+		return (int)(uintptr_t)ptr;
 
 	assert(ptr >= som_buffer && ptr < som_buffer + som_totalSize);
 	//note: we add 1 here, because zero offset is reserved for zero pointer!
@@ -651,7 +651,7 @@ ID_INLINE int idProgram::ScriptObjectMemory_Pack(byte *ptr) {
 }
 ID_INLINE byte *idProgram::ScriptObjectMemory_Unpack(int offset) {
 	if (sizeof(void*) == 4)
-		return (byte*)offset;
+		return (byte*)(uintptr_t)offset;
 
 	assert(offset > 0 && offset <= som_totalSize);
 	return som_buffer + offset - 1;
